Make `include_inherents` an explicit argument
