+++ cumulus/client/cirrus-executor/src/processor.rs
@@ -6,7 +6,6 @@ use sp_runtime::{generic::BlockId, traits::Block as BlockT};
 use std::{
 	collections::{BTreeMap, VecDeque},
 	fmt::Debug,
-	sync::atomic::Ordering,
 };
 
 use cirrus_node_primitives::ProcessorResult;
@@ -73,7 +72,7 @@ where
 {
 	/// Actually implements `process_bundles`.
 	pub(super) async fn process_bundles_impl(
-		self,
+		&self,
 		primary_hash: PHash,
 		bundles: Vec<OpaqueBundle>,
 		shuffling_seed: Randomness,
@@ -135,9 +134,6 @@ where
 
 		let _final_extrinsics =
 			shuffle_extrinsics::<<Block as BlockT>::Extrinsic>(extrinsics, shuffling_seed);
-
-		// Reset the signal to include the inherent extrinsics on next `produce_bundle`.
-		self.to_include_inherents.store(true, Ordering::Relaxed);
 
 		// TODO: now we have the final transaction list:
 		// - apply each tx one by one.
